fix: large matrices (MATELEM)

+++ libpolys/polys/matpol.h
@@ -26,9 +26,9 @@ class ip_smatrix
   #define MATROWS(i) ((i)->nrows)
   #define MATCOLS(i) ((i)->ncols)
   /// 1-based access to matrix
-  #define MATELEM(mat,i,j) ((mat)->m)[MATCOLS((mat)) * ((i)-1) + (j)-1]
+  #define MATELEM(mat,i,j) ((mat)->m)[(long)MATCOLS((mat)) * (long)((i)-1) + (long)(j)-1]
   /// 0-based access to matrix
-  #define MATELEM0(mat,i,j) ((mat)->m)[MATCOLS((mat)) * (i) + (j)]
+  #define MATELEM0(mat,i,j) ((mat)->m)[(long)MATCOLS((mat)) * (long)(i) + (long)(j)]
 };
 
 enum DetVariant
